fixup! efficient struct packing

+++ message/option.go
@@ -127,8 +127,8 @@ const (
 )
 
 type OptionDef struct {
-	MinLen      int
-	MaxLen      int
+	MinLen      uint32
+	MaxLen      uint32
 	ValueFormat ValueFormat
 }
 
@@ -416,7 +416,7 @@ func (o *Option) Unmarshal(data []byte, optionDefs map[OptionID]OptionDef, Optio
 			// Skip unrecognized options (RFC7252 section 5.4.1)
 			return len(data), nil
 		}
-		if len(data) < def.MinLen || len(data) > def.MaxLen {
+		if uint32(len(data)) < def.MinLen || uint32(len(data)) > def.MaxLen {
 			// Skip options with illegal value length (RFC7252 section 5.4.3)
 			return len(data), nil
 		}
